Changed Main.py -> db.py
Changed the name of the database away from main to DB.
This allows the db to be used without being the main function.
To maintain the unit tests added, the if __name__ == "__Main__":
clause has been added which will run the unit tests if the
file is ever run by itself.